fix signup
